feat: expose independent Avatar renderer and editor entries
