+++ packages/react/package.json
@@ -1,10 +1,12 @@
 {
   "name": "@oneworks/avatar-react",
-  "version": "1.0.0-rc.8",
+  "version": "1.0.0-rc.9",
   "description": "React renderer and embeddable editor for OneWorks 3D Avatar",
   "type": "module",
   "license": "MIT",
-  "sideEffects": ["./dist/style.css"],
+  "sideEffects": [
+    "./dist/*.css"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/oneworks-ai/avatar.git",
@@ -20,9 +22,31 @@
       "__oneworks__": "./src/style.scss",
       "default": "./dist/style.css"
     },
-    "./package.json": "./package.json"
+    "./package.json": "./package.json",
+    "./renderer": {
+      "__oneworks__": "./src/renderer.tsx",
+      "types": "./dist/renderer.d.ts",
+      "default": "./dist/renderer.js"
+    },
+    "./renderer.css": {
+      "__oneworks__": "./src/renderer-style.scss",
+      "default": "./dist/renderer.css"
+    },
+    "./editor": {
+      "__oneworks__": "./src/editor.tsx",
+      "types": "./dist/editor.d.ts",
+      "default": "./dist/editor.js"
+    },
+    "./editor.css": {
+      "__oneworks__": "./src/style.scss",
+      "default": "./dist/editor.css"
+    }
   },
-  "files": ["dist", "README.md", "README.zh-Hans.md"],
+  "files": [
+    "dist",
+    "README.md",
+    "README.zh-Hans.md"
+  ],
   "scripts": {
     "build": "vite build",
     "typecheck": "tsc -p tsconfig.json --noEmit --pretty false"
@@ -34,5 +58,7 @@
     "react": ">=18.3.0 <20",
     "react-dom": ">=18.3.0 <20"
   },
-  "publishConfig": { "access": "public" }
+  "publishConfig": {
+    "access": "public"
+  }
 }
