+++ packages/react/src/pickers.scss
@@ -1,0 +1,94 @@
+.oneworks-avatar-picker {
+  display: grid;
+  gap: 8px;
+  min-width: 0;
+  color: var(--text-color, currentColor);
+  font-family:
+    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
+    'Segoe UI', sans-serif;
+}
+
+.oneworks-avatar-picker *,
+.oneworks-avatar-picker *::before,
+.oneworks-avatar-picker *::after {
+  box-sizing: border-box;
+}
+
+.oneworks-avatar-picker__search {
+  width: 100%;
+  min-width: 0;
+  height: 36px;
+  padding: 0 12px;
+  color: inherit;
+  background: var(--sub-bg-color, transparent);
+  border: 1px solid var(--border-color, color-mix(in srgb, currentColor 18%, transparent));
+  border-radius: 8px;
+  outline: none;
+}
+
+.oneworks-avatar-picker__search:focus-visible {
+  border-color: var(--primary-color, #e23f12);
+  box-shadow: 0 0 0 2px color-mix(in srgb, var(--primary-color, #e23f12) 20%, transparent);
+}
+
+.oneworks-avatar-picker__grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, 32px);
+  gap: 8px;
+  align-items: center;
+  min-width: 0;
+}
+
+.oneworks-avatar-picker__option {
+  position: relative;
+  display: grid;
+  place-items: center;
+  width: 32px;
+  height: 32px;
+  min-width: 32px;
+  padding: 0;
+  overflow: hidden;
+  color: inherit;
+  background: transparent;
+  border: 1px solid transparent;
+  border-radius: 8px;
+  cursor: pointer;
+}
+
+.oneworks-avatar-picker__option:hover:not(:disabled) {
+  border-color: var(--sub-border-color, color-mix(in srgb, currentColor 28%, transparent));
+}
+
+.oneworks-avatar-picker__option[data-selected='true'] {
+  border-color: var(--primary-color, #e23f12);
+  box-shadow: 0 0 0 1px var(--primary-color, #e23f12);
+}
+
+.oneworks-avatar-picker__option:focus-visible {
+  outline: 2px solid var(--primary-color, #e23f12);
+  outline-offset: 2px;
+}
+
+.oneworks-avatar-picker__option:disabled {
+  cursor: not-allowed;
+  opacity: .4;
+}
+
+.oneworks-avatar-picker__option > img,
+.oneworks-avatar-picker__avatar-preview {
+  display: block;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.oneworks-avatar-picker__fallback {
+  font-size: 12px;
+  line-height: 1;
+}
+
+.oneworks-avatar-picker__empty {
+  grid-column: 1 / -1;
+  color: var(--sub-text-color, color-mix(in srgb, currentColor 58%, transparent));
+  font-size: 12px;
+}
