+++ packages/react/src/renderer-style.scss
@@ -1,0 +1,3 @@
+// Complete renderer stylesheet without editor UI or global design tokens.
+@use '../../../src/InteractiveAvatar';
+@use './renderer';
