+++ packages/react/src/renderer.scss
@@ -1,0 +1,34 @@
+.oneworks-avatar {
+  box-sizing: border-box;
+  color: var(--text-color);
+  font-family:
+    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
+    'Segoe UI', sans-serif;
+}
+
+.oneworks-avatar * {
+  box-sizing: border-box;
+}
+
+.oneworks-avatar {
+  position: relative;
+  display: block;
+  width: 100%;
+  aspect-ratio: 1;
+  overflow: hidden;
+  background: var(--oneworks-avatar-background, transparent);
+}
+
+.oneworks-avatar[data-frame='rounded'] {
+  border-radius: 12.5%;
+}
+
+.oneworks-avatar[data-frame='circle'] {
+  border-radius: 50%;
+}
+
+.oneworks-avatar > .interactive-avatar {
+  width: 100%;
+  height: 100%;
+}
+
