+++ packages/react/src/style.scss
@@ -1,183 +1,4 @@
-@import '@oneworks/route-layout/design-tokens.css';
-
-.oneworks-avatar,
-.oneworks-avatar-editor {
-  box-sizing: border-box;
-  color: var(--text-color);
-  font-family:
-    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
-    'Segoe UI', sans-serif;
-}
-
-.oneworks-avatar *,
-.oneworks-avatar-editor * {
-  box-sizing: border-box;
-}
-
-.oneworks-avatar {
-  position: relative;
-  display: block;
-  width: 100%;
-  aspect-ratio: 1;
-  overflow: hidden;
-  background: var(--oneworks-avatar-background, transparent);
-}
-
-.oneworks-avatar[data-frame='rounded'] {
-  border-radius: 12.5%;
-}
-
-.oneworks-avatar[data-frame='circle'] {
-  border-radius: 50%;
-}
-
-.oneworks-avatar > .interactive-avatar {
-  width: 100%;
-  height: 100%;
-}
-
-.oneworks-avatar-editor {
-  position: relative;
-  display: block;
-  width: 100%;
-  min-width: 0;
-  min-height: 640px;
-  overflow: hidden;
-  background: var(--bg-color);
-}
-
-.oneworks-avatar-editor > .avatar-app {
-  min-height: inherit;
-}
-
-.oneworks-avatar-editor > .avatar-app.dark {
-  --app-chrome-overlay-background: color-mix(
-    in srgb,
-    var(--bg-color) 42%,
-    transparent
-  );
-  --bg-color: #141414;
-  --sub-bg-color: #262626;
-  --sub-sub-bg-color: #303030;
-  --text-color: #ffffff;
-  --sub-text-color: #8c8c8c;
-  --sub-sub-text-color: #a5a5a5;
-  --border-color: #303030;
-  --sub-border-color: #434343;
-  --sub-sub-border-color: #595959;
-  --placeholder-color: #575859;
-  --tag-bg: #1f1f1f;
-  --tag-hover-bg: #262626;
-  --code-bg: #1e293b;
-  --code-border: #334155;
-  --star-color: #ffb81c;
-  --primary-color: #e23f12;
-  --primary-soft-bg: color-mix(
-    in srgb,
-    var(--primary-color) 12%,
-    var(--bg-color)
-  );
-  --success-color: #2dd4bf;
-  --warning-color: #f59e0b;
-  --primary-text-color: color-mix(
-    in srgb,
-    var(--primary-color) 82%,
-    var(--text-color)
-  );
-  --danger-color: #f87171;
-}
-
-.oneworks-avatar-picker {
-  display: grid;
-  gap: 8px;
-  min-width: 0;
-  color: var(--text-color, currentColor);
-  font-family:
-    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
-    'Segoe UI', sans-serif;
-}
-
-.oneworks-avatar-picker *,
-.oneworks-avatar-picker *::before,
-.oneworks-avatar-picker *::after {
-  box-sizing: border-box;
-}
-
-.oneworks-avatar-picker__search {
-  width: 100%;
-  min-width: 0;
-  height: 36px;
-  padding: 0 12px;
-  color: inherit;
-  background: var(--sub-bg-color, transparent);
-  border: 1px solid var(--border-color, color-mix(in srgb, currentColor 18%, transparent));
-  border-radius: 8px;
-  outline: none;
-}
-
-.oneworks-avatar-picker__search:focus-visible {
-  border-color: var(--primary-color, #e23f12);
-  box-shadow: 0 0 0 2px color-mix(in srgb, var(--primary-color, #e23f12) 20%, transparent);
-}
-
-.oneworks-avatar-picker__grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, 32px);
-  gap: 8px;
-  align-items: center;
-  min-width: 0;
-}
-
-.oneworks-avatar-picker__option {
-  position: relative;
-  display: grid;
-  place-items: center;
-  width: 32px;
-  height: 32px;
-  min-width: 32px;
-  padding: 0;
-  overflow: hidden;
-  color: inherit;
-  background: transparent;
-  border: 1px solid transparent;
-  border-radius: 8px;
-  cursor: pointer;
-}
-
-.oneworks-avatar-picker__option:hover:not(:disabled) {
-  border-color: var(--sub-border-color, color-mix(in srgb, currentColor 28%, transparent));
-}
-
-.oneworks-avatar-picker__option[data-selected='true'] {
-  border-color: var(--primary-color, #e23f12);
-  box-shadow: 0 0 0 1px var(--primary-color, #e23f12);
-}
-
-.oneworks-avatar-picker__option:focus-visible {
-  outline: 2px solid var(--primary-color, #e23f12);
-  outline-offset: 2px;
-}
-
-.oneworks-avatar-picker__option:disabled {
-  cursor: not-allowed;
-  opacity: .4;
-}
-
-.oneworks-avatar-picker__option > img,
-.oneworks-avatar-picker__avatar-preview {
-  display: block;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.oneworks-avatar-picker__fallback {
-  font-size: 12px;
-  line-height: 1;
-}
-
-.oneworks-avatar-picker__empty {
-  grid-column: 1 / -1;
-  color: var(--sub-text-color, color-mix(in srgb, currentColor 58%, transparent));
-  font-size: 12px;
-}
+// Full compatibility stylesheet; renderer.css is the display-only alternative.
+@use './editor';
+@use './renderer';
+@use './pickers';
